WIP Display modifier key names
When user modifies a key binding, three buttons for each modifier are
shown. Label those buttons with key name bound to specific modifier.

In result, by default buttons are labelled:
"Left Ctrl" (instead of "mod1")
"Left Alt" (instead of "mod2")

Do not show the button at all, if the modifier is not bound to any key
(as is case for "mod3" by default).