Disable PVS warning suggesting different container
The problematic warning is V826 "Consider replacing standard container
with a different one".

PVS leaves "helpful" warnings like: "Consider replacing the '<name>'
std::list with std::vector.  Contiguous placement of elements in memory
can be more efficient."

Yes, using std::list is generally a bad idea, but std::vector is *not* a
drop-in replacement.  In some cases refactoring code depending on
std::list behaviour to use std::vector is extremely risky due to
std::vector reinitializing the storage buffer.

One example of such risky and very hard to accompplish refactor is
sdl_mapper - code of which is *not* performance critical, so std::list
can be left as-is.